pahole: Only warn about multithreading not being available with older versions of elfutils in verbose mode
For kernel builds, pahole is called with -j to specify multithreading
as long as it is >= 1.22.

However, if elfutils are older than 0.178 at compile time,
multithreading is not supported and we emit a warning to stderr.

The problem is this message will be emitted each time pahole is run, and
this translates to thousands of times in a kernel build (once for
vmlinux and once for each module).

Often users who are not familiar with BTF generation or pahole
reasonably assume this is an error, whereas it is actually fine; it just
means that pahole is a bit slower to generate BTF.

Move the message to verbose mode only, as in debugging cases it is
important to know if pahole is running in parallel.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Link: https://lore.kernel.org/r/20240801095058.1946790-1-alan.maguire@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>