dwarf_loader: Allow filtering CUs early in loading
With this, we can avoid warnings for unsupported DWARF tags like:

  die__process_function: tag not supported 0x2f (template_type_parameter)!

when processing object files generated from languages such as Rust, for
instance when building the Linux kernel with `CONFIG_RUST`, after
applying the next patch in this series.

As an added bonus, this should speed up processing of large objects with
filtered CUs, as their details will no longer be walked.

Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Signed-off-by: Matthew Maurer <mmaurer@google.com>
Tested-by: Alan Maguire <alan.maguire@oracle.com>
Cc: Alice Ryhl <aliceryhl@google.com>
Cc: rust-for-linux@vger.kernel.org
Link: https://lore.kernel.org/r/20240801185054.2518383-1-mmaurer@google.com
[ Split from a larger patch ]
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>