pahole: Reduce memory usage by smarter deleting of CUs
pahole uses a lot of memory when doing DWARF->BTF encoding.  While some
of this is inevitable there are some cases where we can improve things.

When doing BTF encoding, we sometimes need to keep CUs around to compare
functions to spot optimized/inconsistent functions.  Currently we decide
this on a per-encoder basis (saved_func_cnt) where we count how many
times we save a function, and if any functions are saved for the
_encoder_, we keep all CUs.

Instead we can have a per-CU count of saved functions, and if none are
saved for the _CU_ it is safe to delete it.

This leads to deleting ~90 CUs during parallel vmlinux BTF generation
versus deleting just 1 prior to this change.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Link: https://lore.kernel.org/r/20240826110148.790124-1-alan.maguire@oracle.com
Cc: dwarves@vger.kernel.org
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>