pfunct: Print DW_AT_external=1 functions once
Namhyung mentioned that 'set_task_cpu' wasn't being printed with pfunct,
that was because those functions (with DW_AT_external=1) were not being
printed at all, out of some wrong expectations, this was added to avoid
printing it multiple times to allow for things like --compile to work,
but that was too much of a big hammer.

So add fn_stats->printed and use it to determine if functions were
already printed.

Also don't print definitions for functions marked as declarations, as it
clashes with the ones in system headers, like __builtin_memcpy, etc.

This all needs a more thorough love and care, but at least for now
--compile works with 'fullcircle tcp.o' (generating the functions and
the types it uses and then compiling it to get the original types and
functions and prints the external functions:

  $ pfunct -F dwarf -f set_task_cpu vmlinux
  void set_task_cpu(struct task_struct * p, unsigned int new_cpu);
  $ pfunct -F btf -f set_task_cpu vmlinux
  void set_task_cpu(struct task_struct * p, unsigned int new_cpu);
  $

Reported-by: Namhyung Kim <namhyung@kernel.org>
Tested-by: Namhyung Kim <namhyung@kernel.org>
Cc: Alan Maguire <alan.maguire@oracle.com>
Cc: Jiri Olsa <jolsa@kernel.org>
Link: https://lore.kernel.org/all/Zs0QVVZZfN9H7wQC@x1
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>