pfunct: Show all functions that match filter criteria
Currently, "pfunct --function <name> vmlinux" will bail once it finds
an instance of a function; add an option

  -A, --all                  show all functions that match filter, or show all
                             function prototypes if no filter is specified

..that allows us to see all instances of a function.  It will be
useful in testing functions skipped by BTF encoding.

When --function is not specified, all prototypes are displayed but
we make use of the pfunct_stealer for this case as it massively
speeds up processing from ~6min to ~12sec.

It will be useful in tests to be able to collect prototypes quickly.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Tested-by: Arnaldo Carvalho de Melo <acme@redhat.com>
Cc: Eduard Zingerman <eddyz87@gmail.com>
Cc: Jiri Olsa <jolsa@kernel.org>
Cc: Jiri Slaby <jirislaby@kernel.org>
Cc: Shung-Hsi Yu <shung-hsi.yu@suse.com>
Link: https://lore.kernel.org/r/20240916134946.3893204-3-alan.maguire@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>