tests/reproducible_build: Fix vmlinux finding logic
In updating the logic that finds the path for vmlinux, it is useful to
support a few different options

1. command-line parameter specification
2. retrieval via pahole option
3. environment variable based retrieval

Options 2 and 3 are useful for the case where the script is run via the
tests script.  When adding support for 3 I incorrectly specified

vmlinux=${vmlinux:-1}

...which should have been

vmlinux=${vmlinux:-$1}

i.e. assign var vmlinux to the env var vmlinux if set, and if that is
not set, use the first parameter for the vmlinux value.  If neither are
set we can then fall back to the pahole option.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Link: https://lore.kernel.org/r/20241002113326.792495-2-alan.maguire@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>