btf_encoder: Use bitfield to control var encoding
We will need more granularity in the future, in order to add support for
encoding global variables as well.

So replace the skip_encoding_vars boolean with a flag variable named
"encode_vars".

There is currently only one bit specified, and it is set when percpu
variables should be emitted.

Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Link: https://lore.kernel.org/r/20241002235253.487251-2-stephen.s.brennan@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>