btf_encoder: Stop indexing symbols for VARs
Currently we index symbols from the percpu ELF section, and when
processing DWARF variables for inclusion, we check whether the variable
matches an existing symbol. The matched symbol is used for three
purposes:

1. When no symbol of the same address is found, the variable is skipped.
   This can occur because the symbol name was an invalid BTF
   identifier, and so it did not get indexed. Or more commonly, it can
   be because the variable is not stored in the per-cpu section, and
   thus was not indexed.

2. If the symbol offset is 0, then we compare the DWARF variable's name
   against the symbol name to filter out "special" DWARF variables.

3. We use the symbol size in the DATASEC entry for the variable.

For 1, we don't need the symbol table: we can simply check the DWARF
variable name directly, and we can use the variable address to determine
the ELF section it is contained in.

For 3, we also don't need the symbol table: we can use the variable's
size information from DWARF.

Issue 2 is more complicated, but thanks to the addition of the
"artificial" and "top_level" flags, many of the "special" DWARF
variables can be directly filtered out, and the few remaining
problematic variables can be filtered by name from a kernel-specific
list of patterns.

This allows the symbol table index to be removed. The benefit of
removing this index is twofold. First, handling variable addresses is
simplified, since we don't need to know whether the file is ET_REL.

Second, this will make it easier to output variables that aren't just
percpu, since we won't need to index variables from all ELF sections.

Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Acked-by: Jiri Olsa <jolsa@kernel.org>
Link: https://lore.kernel.org/r/20241004172631.629870-2-stephen.s.brennan@oracle.com
Link: https://lore.kernel.org/r/20241005000147.723515-1-stephen.s.brennan@oracle.com
[ This includes the fixup in the second Link:, from Stephen ]
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>