btf_encoder: Explicitly check addr/size for u32 overflow
The addr is a uint64_t, and depending on the size of a data section,
there's no guarantee that it fits into a uint32_t, even after
subtracting out the section start address. Similarly, the variable size
is a size_t which could exceed a uint32_t.

Check both for overflow, and if found, skip the variable with an error
message. Use explicit casts when we cast to uint32_t so it's plain to
see that this is happening.

Reviewed-by: Alan Maguire <alan.maguire@oracle.com>
Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Acked-by: Jiri Olsa <jolsa@kernel.org>
Link: https://lore.kernel.org/r/20241004172631.629870-3-stephen.s.brennan@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>