dwarf_loader: Use libdw__lock for dwarf_getlocation(s)
Eduard noticed [1] intermittent segmentation faults triggered by caching
done internally in dwarf_getlocation(s).

A binary tree of location information is cached in the CU, and if
multiple threads access it concurrently we can get segmentation faults.

It is possible to compile elfutils with experimental thread-safe
support, but safer for now to add locking to pahole.

No additional overhead in pahole encoding was observed
as a result of these changes.

[1] https://lore.kernel.org/dwarves/080794545d8eb3df3d6eba90ac621111ab7171f5.camel@gmail.com/

Reported-by: Eduard Zingerman <eddyz87@gmail.com>
Suggested-by: Arnaldo Carvalho de Melo <acme@kernel.org>
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>
Tested-by: Jiri Olsa <jolsa@kernel.org>
Acked-by: Eduard Zingerman <eddyz87@gmail.com>
Acked-by: Jiri Olsa <jolsa@kernel.org>
Cc: Alexei Starovoitov <ast@kernel.org>
Cc: Andrii Nakryiko <andrii@kernel.org>
Cc: Daniel Borkmann <daniel@iogearbox.net>
Cc: Song Liu <song@kernel.org>
Cc: Yonghong Song <yonghong.song@linux.dev>
Link: https://lore.kernel.org/r/20241115113605.1504796-3-alan.maguire@oracle.com
Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>