btf_encoder: Filter out __gendwarfksyms_ptr_
With CONFIG_GENDWARFKSYMS, __gendwarfksyms_ptr_<symbol>
variables are added to the kernel in EXPORT_SYMBOL() to ensure
DWARF type information is available for exported symbols in the
TUs where they're actually exported. These symbols are dropped
when linking vmlinux, but dangling references to them remain
in DWARF, which results in thousands of 0 address variables
that pahole needs to validate (since commit 9810758003ce
("btf_encoder: Verify 0 address DWARF variables are in ELF
section")).

Filter out symbols with the __gendwarfksyms_ptr_ name prefix in
filter_variable_name() instead of calling variable_in_sec()
for all of them. This reduces the time it takes to process
.tmp_vmlinux1 by ~77% on my test system:

Before: 35.775 +- 0.121  seconds time elapsed  ( +-  0.34% )
 After: 8.3516 +- 0.0407 seconds time elapsed  ( +-  0.49% )

Fixes: 9810758003ce9f ("btf_encoder: Verify 0 address DWARF variables
are in ELF section")

Signed-off-by: Sami Tolvanen <samitolvanen@google.com>
Acked-by: Jiri Olsa <jolsa@kernel.org>
Cc: Arnaldo Carvalho de Melo <acme@kernel.org>
Cc: Yonghong Song <yonghong.song@linux.dev>
Cc: Alexei Starovoitov <ast@kernel.org>
Cc: Andrii Nakryiko <andrii@kernel.org>
Cc: Daniel Borkmann <daniel@iogearbox.net>
Cc: Song Liu <song@kernel.org>
Cc: Eduard Zingerman <eddyz87@gmail.com>
Cc: Stephen Brennan <stephen.s.brennan@oracle.com>
Cc: Laura Nao <laura.nao@collabora.com>
Cc: Uros Bizjak <ubizjak@gmail.com>
Cc: Cong Wang <cong.wang@bytedance.com>
Link: https://lore.kernel.org/dwarves/20250317222424.3837495-1-samitolvanen@google.com/
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>