dwarf_loader: Fix termination on BTF encoding error
When BTF encoding thread aborts because of an error, dwarf loader
worker threads get stuck in cus_queue__enqdeq_job() at:

    pthread_cond_wait(&cus_processing_queue.job_added, &cus_processing_queue.mutex);

To avoid this, introduce an abort flag into cus_processing_queue, and
atomically check for it in the deq loop. The flag is only set in case
of a worker thread exiting on error. Make sure to pthread_cond_signal
to the waiting threads to let them exit too.

In cus__process_file fix the check of an error returned from
dwfl_getmodules: it may return a positive number when a
callback (cus__process_dwflmod in our case) returns an error.

Link: https://lore.kernel.org/dwarves/Z-JzFrXaopQCYd6h@localhost/

Reported-by: Domenico Andreoli <domenico.andreoli@linux.com>
Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>
Tested-by: Domenico Andreoli <domenico.andreoli@linux.com>
Cc: Arnaldo Carvalho de Melo <acme@kernel.org>
Cc: Andrii Nakryiko <andrii@kernel.org>
Cc: Eduard Zingerman <eddyz87@gmail.com>
Cc: Mykola Lysenko <mykolal@fb.com>
Link: https://lore.kernel.org/dwarves/20250328174003.3945581-1-ihor.solodrai@linux.dev/
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>