btf_encoder: Skip functions consuming packed structs passed by value on stack
Most ABIs allow functions to receive structs passed by value, if they
fit in a register or a pair of registers, depending on the exact ABI.
However, when there is a struct passed by value but all registers are
already used for parameters passing, the struct is still passed by value
but on the stack. This becomes an issue if the passed struct is defined
with some attributes like __attribute__((packed)) or
__attribute__((aligned(X)), as its location on the stack is altered, but
this change is not reflected in dwarf information. The corresponding BTF
data generated from this can lead to incorrect BPF trampolines
generation (eg to attach bpf tracing programs to kernel functions) in
the Linux kernel.

Prevent those wrong cases by not encoding functions consuming structs
passed by value on stack, when those structs do not have the expected
alignment due to some attribute usage.

Signed-off-by: Alexis Lothoré (eBPF Foundation) <alexis.lothore@bootlin.com>
Cc: Arnaldo Carvalho de Melo <acme@kernel.org>
Cc: Alexei Starovoitov <ast@fb.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Cc: Bastien Curutchet <bastien.curutchet@bootlin.com>
Cc: ebpf@linuxfoundation.org
Acked-by: Ihor Solodrai <ihor.solodrai@linux.dev>
Acked-by: Jiri Olsa <jolsa@kernel.org>
Link: https://lore.kernel.org/dwarves/20250707-btf_skip_structs_on_stack-v3-1-29569e086c12@bootlin.com/
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>