btf_encoder: group all function ELF syms by function name
btf_encoder collects function ELF symbols into a table, which is later
used for processing DWARF data and determining whether a function can
be added to BTF.

So far the ELF symbol name was used as a key for search in this table,
and a search by prefix match was attempted in cases when ELF symbol
name has a compiler-generated suffix.

This implementation has bugs [1][2], causing some functions to be
inappropriately excluded from (or included into) BTF.

Rework the implementation of the ELF functions table. Use a name of a
function without any suffix - a symbol name before the first
occurrence of '.' - as a key. This way btf_encoder__find_function()
always returns a valid elf_function object (or NULL).

Collect an array of symbol name + address pairs from GElf_Sym for each
elf_function when building the elf_functions table.

Take into account that some symbols associated with a function name
are not relevant, and do not collect them into elf_functions table:
  * ".cold" suffix indicates a piece of hot/cold split
  * ".part" suffix indicates a piece of partial inline

When inspecting symbol name we have to search for any occurrence of
the target suffix, as opposed to testing the entire suffix, or the end
of a string. This is because suffixes may be combined by the compiler,
for example producing ".isra0.cold", and the conclusion will be
incorrect.

Introduce ambiguous_addr flag to the elf_function. It is set when the
functions are deduped in elf_functions__collect() examining the array
of ELF symbols in elf_function__has_ambiguous_address(). It tests
whether there is only one unique address for this function name.

In btf_encoder__add_saved_funcs() check ambiguous_addr flag when
deciding whether a function should be included in BTF.

[1] https://lore.kernel.org/bpf/2f8c792e-9675-4385-b1cb-10266c72bd45@linux.dev/
[2] https://lore.kernel.org/dwarves/6b4fda90fbf8f6aeeb2732bbfb6e81ba5669e2f3@linux.dev/

v2->v3:
  * Move ambiguous_addr flag from btf_encoder_func_state to
    elf_function. We don't have to process DWARF to know the value of
    ambiguous_addr, so elf_function is more appropriate.
  * Don't even add symbols with non_fn suffixes into elf_functions
    table. This saves us space and string comparisons. (Alan)
  * Increase elf_function->sym_cnt size from u8 to u16, and other
    nits. (Jiri)

v2: https://github.com/acmel/dwarves/pull/68/commits/f23968d964eb50359715257962a6f9e07c8cf793
v1: https://lore.kernel.org/dwarves/98f41eaf6dd364745013650d58c5f254a592221c@linux.dev/

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>